Support EntityFallPacket
Support EntityFallPacket.
